osx build mods

+++ src/rpcblockchain.cpp
@@ -46,7 +46,7 @@ double GetDifficultyINTERNAL(const CBlockIndex* blockindex, bool networkDifficul
     int nShiftAmount = (powLimit >> 24) & 0xff;
 
     double dDiff =
-        (double)(powLimit & 0x00ffffff) / 
+        (double)(powLimit & 0x00ffffff) /
         (double)(bits & 0x00ffffff);
 
     while (nShift < nShiftAmount)
@@ -474,14 +474,14 @@ Value minerids(const Array& params, bool fHelp)
                 for (j=0; j<33; j++)
                     sprintf(&hexstr[j*2],"%02x",pubkeys[i][j]);
                 item.push_back(Pair("notaryid", i));
-                
+
                 bitcoin_address(kmdaddr,60,pubkeys[i],33);
                 m = (int32_t)strlen(kmdaddr);
                 kmdaddress.resize(m);
                 ptr = (char *)kmdaddress.data();
                 memcpy(ptr,kmdaddr,m);
                 item.push_back(Pair("KMDaddress", kmdaddress));
-                
+
                 item.push_back(Pair("pubkey", hex));
                 item.push_back(Pair("blocks", tally[i]));
                 a.push_back(item);
@@ -639,7 +639,7 @@ Value paxprices(const Array& params, bool fHelp)
         else
         {
             CBlockIndex *pblockindex = chainActive[heights[i]];
-            
+
             item.push_back(Pair("t", (int64_t)pblockindex->nTime));
             item.push_back(Pair("p", (double)prices[i] / COIN));
             a.push_back(item);
@@ -840,7 +840,7 @@ Value getblockchaininfo(const Array& params, bool fHelp)
 
     ZCIncrementalMerkleTree tree;
     pcoinsTip->GetAnchorAt(pcoinsTip->GetBestAnchor(), tree);
-    obj.push_back(Pair("commitments",           tree.size()));
+    obj.push_back(Pair("commitments",           (uint64_t)tree.size()));
 
     const Consensus::Params& consensusParams = Params().GetConsensus();
     CBlockIndex* tip = chainActive.Tip();
